+++ scripts/macros/2024/speciesFeatures/aasimar/celestialRevelation.js
@@ -1,4 +1,4 @@
-import {activityUtils, actorUtils, effectUtils, genericUtils, itemUtils, tokenUtils, workflowUtils} from '../../../../utils.js';
+import {activityUtils, actorUtils, combatUtils, dialogUtils, effectUtils, genericUtils, itemUtils, tokenUtils, workflowUtils} from '../../../../utils.js';
 async function use({trigger, workflow}) {
     let identifier = activityUtils.getIdentifier(workflow.activity);
     if (!identifier) return;
@@ -7,6 +7,8 @@ async function use({trigger, workflow}) {
     let effectData = genericUtils.duplicate(sourceEffect.toObject());
     effectData.duration = itemUtils.convertDuration(workflow.activity);
     effectData.origin = sourceEffect.uuid;
+    let damageType = identifier === 'necroticShroud' ? 'necrotic' : 'radiant';
+    genericUtils.setProperty(effectData, 'flags.chris-premades.celestialRevelation.damageType', damageType);
     let avatarImg = itemUtils.getConfig(workflow.item, identifier + 'Avatar');
     let tokenImg = itemUtils.getConfig(workflow.item, identifier + 'Token');
     let avatarImgPriority = itemUtils.getConfig(workflow.item, 'avatarPriority');
@@ -29,6 +31,28 @@ async function turnEnd({trigger: {entity: effect}}) {
     let nearby = tokenUtils.findNearby(token, 10, 'all', {includeIncapacitated: true});
     if (!nearby.length) return;
     await workflowUtils.syntheticActivityRoll(activity, nearby);
+}
+async function damage({trigger: {entity: effect}, workflow}) {
+    if (!workflow.targets.size || !(workflow.item.type === 'spell' || workflowUtils.isAttackType(workflow, 'attack'))) return;
+    if (!combatUtils.perTurnCheck(effect, 'celestialRevelation', true, workflow.token.id)) return;
+    let damageType = effect.flags['chris-premades']?.celestialRevelation?.damageType;
+    if (!damageType) return;
+    if (workflow.targets.size === 1) {
+        let selection = await dialogUtils.confirmUseItem(effect);
+        if (!selection) return;
+        await workflowUtils.bonusDamage(workflow, String(workflow.actor.system.attributes.prof), {damageType});
+    } else {
+        let result = await dialogUtils.selectTargetDialog(effect.name, genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: effect.name}), workflow.targets, {skipDeadAndUnconscious: false});
+        if (!result) return;
+        genericUtils.setProperty(workflow, 'chris-premades.celestialRevelation', {target: result[0].document.uuid, damage: workflow.actor.system.attributes.prof, damageType});
+    }
+    await combatUtils.setTurnCheck(effect, 'celestialRevelation');
+}
+async function applyDamage({trigger, workflow, ditem}) {
+    if (!workflow['chris-premades']?.celestialRevelation) return;
+    let {target, damage, damageType} = workflow['chris-premades'].celestialRevelation;
+    if (target !== ditem.targetUuid || !damage) return;
+    workflowUtils.modifyDamageAppliedFlat(ditem, damage, {type: damageType, multiplier: 'auto'});
 }
 export let celestialRevelation = {
     name: 'Celestial Revelation',
@@ -107,6 +131,20 @@ export let celestialRevelationInnerRadiance = {
     name: 'Celestial Revelation: Inner Radiance',
     version: celestialRevelation.version,
     rules: celestialRevelation.rules,
+    midi: {
+        actor: [
+            {
+                pass: 'damageRollComplete',
+                macro: damage,
+                priority: 250
+            },
+            {
+                pass: 'applyDamage',
+                macro: applyDamage,
+                priority: 250
+            }
+        ]
+    },
     combat: [
         {
             pass: 'turnEnd',
@@ -114,4 +152,42 @@ export let celestialRevelationInnerRadiance = {
             priority: 50
         }
     ]
+};
+export let celestialRevelationHeavenlyWings = {
+    name: 'Celestial Revelation: Heavenly Wings',
+    version: celestialRevelation.version,
+    rules: celestialRevelation.rules,
+    midi: {
+        actor: [
+            {
+                pass: 'damageRollComplete',
+                macro: damage,
+                priority: 250
+            },
+            {
+                pass: 'applyDamage',
+                macro: applyDamage,
+                priority: 250
+            }
+        ]
+    }
+};
+export let celestialRevelationNecroticShroud = {
+    name: 'Celestial Revelation: Necrotic Shroud',
+    version: celestialRevelation.version,
+    rules: celestialRevelation.rules,
+    midi: {
+        actor: [
+            {
+                pass: 'damageRollComplete',
+                macro: damage,
+                priority: 250
+            },
+            {
+                pass: 'applyDamage',
+                macro: applyDamage,
+                priority: 250
+            }
+        ]
+    }
 };